% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Check Data}
\usage{
checkData(x)
}
\arguments{
\item{x}{a `tibble` of formatted data that has been created using `formatExport`}
}
\value{
a `tibble`
}
\description{
Flag up any potential erorrs
}
