% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatExport.R
\name{formatExport}
\alias{formatExport}
\title{Format Recruitment Export}
\usage{
formatExport(x)
}
\arguments{
\item{x}{a `tibble` of the raw Jisc output using `readr::read_csv`}
}
\value{
a `tibble`
}
\description{
The recruitment database is populated using a Jisc survey which is e-mailed out to potential contacts. This function
formats the direct export from Jisc ready for validation and upload using `waruDB`
}
