% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_code.R
\name{hash_code}
\alias{hash_code}
\title{Hash Code}
\usage{
hash_code(x)
}
\arguments{
\item{x}{a `tibble` of formatted data}
}
\value{
a numeric vector of distance in km
}
\description{
Create a simple hash code for sanity checking of database
}
\keyword{internal}
