% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_distance.R
\name{location_distance}
\alias{location_distance}
\title{Convert PostCode to distance}
\usage{
location_distance(x)
}
\arguments{
\item{x}{a \code{tibble} of formatted data}
}
\value{
a numeric vector of distance in km
}
\description{
Convert PostCodes to the distance (km) from Wellbeing-and-health Assessment Research Unit (WARU)
}
\keyword{internal}
