% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gender.R
\name{plot_gender}
\alias{plot_gender}
\title{Plot Gender}
\usage{
plot_gender(x)
}
\arguments{
\item{x}{a `tibble` of formatted data which has been created using `formatExport``}
}
\value{
a `ggplot` object
}
\description{
Produce a simple bar plot of the number of records by Gender
}
