% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_location.R
\name{plot_location}
\alias{plot_location}
\title{Plot Location}
\usage{
plot_location(x)
}
\arguments{
\item{x}{a `tibble` of formatted data which has been created using `formatExport``}
}
\value{
a `leaflet` map object
}
\description{
Plot locations on interactive map
}
